DataValidator: use 32bit timeout instead of 64bit, saves memory

+++ validation/data_validator_group.cpp
@@ -64,7 +64,7 @@ DataValidatorGroup::~DataValidatorGroup()
 }
 
 void
-DataValidatorGroup::set_timeout(uint64_t timeout_interval_us)
+DataValidatorGroup::set_timeout(uint32_t timeout_interval_us)
 {
 	DataValidator *next = _first;
